Move S3Client to s3util.py from s3.py
This is to make the import of S3Client from s3op (upcoming change) as cheap
as possible in terms of import time
